requirements.txt
protobuf~=5.26.1
aiohttp~=3.9.5
pydantic-settings~=2.2.1
grpcio
boto3~=1.34.99
uuid~=1.30
blake3~=0.4.1

pydantic~=2.7.1